requirements.txt
simple-salesforce>=1.12.5
pandas>=2.0.0
python-dotenv>=1.0.0
